adding glenn posts!
This will add Glenn posts - we have a current workaround to
remove images from blogger (that cannot be embedded) and can try
to improve upon this in the future. This also ensures that long
(and strange) urls are parsed into something more human readable,
and we add a nice author tag to the front page. It is not perfect,
but like many things, a WIP that will get better

Signed-off-by: vsoch <vsoch@users.noreply.github.com>